Move /react helper functions into utils module
