+++ src/controllers/dev/control/react.command.ts
@@ -1,14 +1,14 @@
-import {
-  ChatInputCommandInteraction,
-  Message,
-  SlashCommandBuilder,
-} from "discord.js";
+import { SlashCommandBuilder } from "discord.js";
 
 import {
   RoleLevel,
   checkPrivilege,
 } from "../../../middleware/privilege.middleware";
 import { CommandBuilder } from "../../../types/command.types";
+import {
+  fetchMessageByIdentifier,
+  fetchMostRecentMessage,
+} from "./dev-control-utils";
 
 const devReact = new CommandBuilder();
 
@@ -55,44 +55,6 @@ devReact.execute(async interaction => {
   await interaction.reply({ content: "👍", ephemeral: true });
   return true;
 });
-
-async function fetchMessageByIdentifier(
-  messageIdentifier: string,
-  interaction: ChatInputCommandInteraction,
-): Promise<Message | null> {
-  const messageId = extractMessageID(messageIdentifier);
-  if (messageId === null) {
-    await interaction.reply({
-      content: `\`${messageIdentifier}\` does not point to a valid message!`,
-      ephemeral: true,
-    });
-    return null;
-  }
-  const message = await interaction.channel!.messages.fetch(messageId);
-  return message;
-}
-
-async function fetchMostRecentMessage(
-  interaction: ChatInputCommandInteraction,
-): Promise<Message> {
-  const messages = await interaction.channel!.messages.fetch({ limit: 1 });
-  const message = Array.from(messages.values())[0];
-  return message;
-}
-
-function extractMessageID(idOrUrl: string): string | null {
-  const idRegexp = /^\d+$/i;
-  // The string is an ID itself.
-  if (idOrUrl.match(idRegexp)) return idOrUrl;
-
-  // Otherwise see if it's a URL. Specifically, we only care about messages
-  // within guild channels at the moment, hence /channels/.
-  const urlRegexp = /^https:\/\/discord\.com\/channels\/\d+\/\d+\/(\d+)$/i;
-  const match = idOrUrl.match(urlRegexp);
-  if (!match) return null;
-  const [, messageId] = match;
-  return messageId;
-}
 
 const devReactSpec = devReact.toSpec();
 export default devReactSpec;
